불필요한 컴포넌트 병합
